+++ libavformat/tls.h
@@ -1,7 +1,8 @@
 /*
- * TLS/SSL Protocol
+ * TLS/DTLS/SSL Protocol
  * Copyright (c) 2011 Martin Storsjo
- *
+ * Copyright (c) 2025 Jack Lau
+ * 
  * This file is part of FFmpeg.
  *
  * FFmpeg is free software; you can redistribute it and/or
